// src/models/lista-posts-viewmodel.ts
export class ListaPostsViewModel {
    ID: number;
    post_author: number;
    post_date: string;
    post_content: string;
    post_title: string;
    post_name: number;
    guid: string;
    meta_value: number;
    id_post_img: number;
    link_imagem: string;
    id_post_categoria: number;
    category: string;
    id_author: number;
    author: string;
}